requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
et-xmlfile==1.1.0
idna==3.7
iniconfig==2.0.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
paho-mqtt==1.6.1
pandas==2.2.2
pluggy==1.4.0
pytest==8.1.1
pytest-dependency==0.6.0
pytest-ordering==0.6
pytest-rerunfailures==14.0
pytest-sugar==1.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
termcolor==2.4.0
tzdata==2024.1
urllib3==2.2.1
